♻️ refactor(monitoring): extract api key summary panel
Move API key summary table rendering, key metrics, and expanded details into a dedicated monitoring component.

The monitoring page now delegates API key presentation while retaining pagination and expansion behavior; validated with type-check, targeted tests, and lint.
