+++ src/features/monitoring/components/ApiKeySummaryPanel.tsx
@@ -1,0 +1,304 @@
+import { Fragment, type ReactNode } from 'react';
+import type { TFunction } from 'i18next';
+import { IconChevronDown, IconChevronUp, IconInfo, IconKey } from '@/components/ui/icons';
+import type { MonitoringApiKeyRow } from '@/features/monitoring/hooks/useMonitoringData';
+import { formatCompactNumber, formatUsd } from '@/utils/usage';
+import { AccountModelUsageTable, AccountTokenMetricGrid } from './AccountOverviewCard';
+import { MonitoringPanel } from './MonitoringPanel';
+import { PaginationControls } from './MonitoringShared';
+import type { AccountSummaryMetric } from './accountOverviewPresentation';
+import styles from '../MonitoringCenterPage.module.scss';
+
+type ApiKeyOverviewColumn = {
+  key: string;
+  label: string;
+};
+
+type ApiKeyPaginationState = {
+  currentPage: number;
+  totalPages: number;
+  pageItems: MonitoringApiKeyRow[];
+  startItem: number;
+  endItem: number;
+};
+
+type ApiKeySummaryPanelProps = {
+  rows: MonitoringApiKeyRow[];
+  columns: ApiKeyOverviewColumn[];
+  pagination: ApiKeyPaginationState;
+  expandedApiKeys: Record<string, boolean>;
+  hasPrices: boolean;
+  locale: string;
+  pageSize: number;
+  pageSizeOptions: readonly number[];
+  emptyState: ReactNode;
+  t: TFunction;
+  onToggleApiKey: (apiKeyId: string) => void;
+  onPageChange: (page: number) => void;
+  onPageSizeChange: (pageSize: number) => void;
+};
+
+const joinShort = (values: string[], limit = 2) => {
+  if (values.length <= limit) {
+    return values.join(', ');
+  }
+  return `${values.slice(0, limit).join(', ')} +${values.length - limit}`;
+};
+
+const buildApiKeySecondaryText = (row: MonitoringApiKeyRow) => {
+  if (row.isUnknown) {
+    if (row.authLabels.length > 0) {
+      return joinShort(row.authLabels, 2);
+    }
+    if (row.sourceLabels.length > 0) {
+      return joinShort(row.sourceLabels, 2);
+    }
+    if (row.channels.length > 0) {
+      return joinShort(row.channels, 2);
+    }
+  }
+  if (row.apiKeyLabel && row.apiKeyMasked && row.apiKeyLabel !== row.apiKeyMasked) {
+    return row.apiKeyMasked;
+  }
+  if (row.apiKeyHash) {
+    return `sha256:${row.apiKeyHash.slice(0, 12)}`;
+  }
+  return '';
+};
+
+const buildApiKeySummaryMetrics = (
+  row: MonitoringApiKeyRow,
+  hasPrices: boolean,
+  locale: string,
+  t: TFunction
+): AccountSummaryMetric[] => [
+  {
+    key: 'total-calls',
+    label: t('monitoring.total_calls'),
+    value: formatCompactNumber(row.totalCalls),
+  },
+  {
+    key: 'success-calls',
+    label: t('monitoring.success_calls'),
+    value: formatCompactNumber(row.successCalls),
+    valueClassName: styles.goodText,
+  },
+  {
+    key: 'failure-calls',
+    label: t('monitoring.failure_calls'),
+    value: formatCompactNumber(row.failureCalls),
+    valueClassName: row.failureCalls > 0 ? styles.badText : undefined,
+  },
+  {
+    key: 'total-tokens',
+    label: t('monitoring.total_tokens'),
+    value: formatCompactNumber(row.totalTokens),
+  },
+  {
+    key: 'input-tokens',
+    label: t('monitoring.input_tokens'),
+    value: formatCompactNumber(row.inputTokens),
+  },
+  {
+    key: 'output-tokens',
+    label: t('monitoring.output_tokens'),
+    value: formatCompactNumber(row.outputTokens),
+  },
+  {
+    key: 'cached-tokens',
+    label: t('monitoring.cached_tokens'),
+    value: formatCompactNumber(row.cachedTokens),
+  },
+  {
+    key: 'estimated-cost',
+    label: t('monitoring.estimated_cost'),
+    value: hasPrices ? formatUsd(row.totalCost) : '--',
+  },
+  {
+    key: 'latest-request-time',
+    label: t('monitoring.latest_request_time'),
+    value: new Date(row.lastSeenAt).toLocaleString(locale),
+  },
+];
+
+function ApiKeySummaryPrimary({
+  row,
+  expanded,
+  onToggle,
+  t,
+}: {
+  row: MonitoringApiKeyRow;
+  expanded: boolean;
+  onToggle: () => void;
+  t: TFunction;
+}) {
+  const secondaryText = buildApiKeySecondaryText(row);
+  const keyLabel = row.isUnknown
+    ? t('monitoring.api_key_unknown_label')
+    : row.apiKeyLabel || row.apiKeyMasked || t('monitoring.api_key_unknown_label');
+
+  return (
+    <button
+      type="button"
+      className={[styles.accountButton, expanded ? styles.expandedAccountButton : '']
+        .filter(Boolean)
+        .join(' ')}
+      onClick={onToggle}
+      aria-expanded={expanded}
+      title={keyLabel}
+    >
+      <span className={styles.accountExpandGlyph} aria-hidden="true">
+        {expanded ? <IconChevronUp size={15} /> : <IconChevronDown size={15} />}
+      </span>
+      <span className={styles.accountIdentityLine}>
+        <span className={styles.apiKeyIcon} aria-hidden="true">
+          <IconKey size={13} />
+        </span>
+        <span className={styles.accountButtonLabel}>{keyLabel}</span>
+      </span>
+      {secondaryText ? <small>{secondaryText}</small> : null}
+    </button>
+  );
+}
+
+function ApiKeyExpandedDetails({
+  row,
+  hasPrices,
+  locale,
+  t,
+}: {
+  row: MonitoringApiKeyRow;
+  hasPrices: boolean;
+  locale: string;
+  t: TFunction;
+}) {
+  const summaryMetrics = buildApiKeySummaryMetrics(row, hasPrices, locale, t);
+
+  return (
+    <div className={styles.apiKeyExpandedDetails}>
+      <div className={styles.accountStructureModelPanel}>
+        <AccountTokenMetricGrid metrics={summaryMetrics} t={t} variant="table" />
+        <AccountModelUsageTable row={row} hasPrices={hasPrices} locale={locale} t={t} />
+      </div>
+    </div>
+  );
+}
+
+export function ApiKeySummaryPanel({
+  rows,
+  columns,
+  pagination,
+  expandedApiKeys,
+  hasPrices,
+  locale,
+  pageSize,
+  pageSizeOptions,
+  emptyState,
+  t,
+  onToggleApiKey,
+  onPageChange,
+  onPageSizeChange,
+}: ApiKeySummaryPanelProps) {
+  return (
+    <MonitoringPanel
+      title={
+        <span className={styles.panelTitleWithHint}>
+          {t('monitoring.api_key_summary_title')}
+          <span title={t('monitoring.api_key_summary_description')}>
+            <IconInfo
+              size={14}
+              className={styles.panelTitleHintIcon}
+              aria-label={t('monitoring.api_key_summary_description')}
+            />
+          </span>
+        </span>
+      }
+      subtitle={t('monitoring.api_key_summary_desc')}
+      className={styles.apiKeyPanel}
+      extra={
+        <div className={styles.inlineMetrics}>
+          <span>{t('monitoring.api_key_summary_keys_count', { count: rows.length })}</span>
+        </div>
+      }
+    >
+      <div className={`${styles.tableWrapper} ${styles.apiKeySummaryTableWrapper}`}>
+        <table className={`${styles.table} ${styles.apiKeySummaryTable}`}>
+          <colgroup>
+            {columns.map((column) => (
+              <col key={column.key} />
+            ))}
+          </colgroup>
+          <thead>
+            <tr>
+              {columns.map((column) => (
+                <th key={column.key}>{column.label}</th>
+              ))}
+            </tr>
+          </thead>
+          <tbody>
+            {pagination.pageItems.map((row) => {
+              const isExpanded = Boolean(expandedApiKeys[row.id]);
+              const keyMetrics = buildApiKeySummaryMetrics(row, hasPrices, locale, t);
+              const keyMetricByKey = new Map(keyMetrics.map((metric) => [metric.key, metric]));
+              const rowClassName = [
+                styles.apiKeySummaryRow,
+                isExpanded ? styles.apiKeySummaryRowExpanded : '',
+              ]
+                .filter(Boolean)
+                .join(' ');
+
+              return (
+                <Fragment key={row.id}>
+                  <tr className={rowClassName}>
+                    <td>
+                      <ApiKeySummaryPrimary
+                        row={row}
+                        expanded={isExpanded}
+                        onToggle={() => onToggleApiKey(row.id)}
+                        t={t}
+                      />
+                    </td>
+                    <td>{keyMetricByKey.get('total-calls')?.value ?? '--'}</td>
+                    <td className={keyMetricByKey.get('success-calls')?.valueClassName}>
+                      {keyMetricByKey.get('success-calls')?.value ?? '--'}
+                    </td>
+                    <td className={keyMetricByKey.get('failure-calls')?.valueClassName}>
+                      {keyMetricByKey.get('failure-calls')?.value ?? '--'}
+                    </td>
+                    <td>{keyMetricByKey.get('total-tokens')?.value ?? '--'}</td>
+                    <td>{keyMetricByKey.get('estimated-cost')?.value ?? '--'}</td>
+                    <td>{keyMetricByKey.get('latest-request-time')?.value ?? '--'}</td>
+                  </tr>
+                  {isExpanded ? (
+                    <tr className={styles.apiKeyDetailRow}>
+                      <td colSpan={columns.length}>
+                        <ApiKeyExpandedDetails row={row} hasPrices={hasPrices} locale={locale} t={t} />
+                      </td>
+                    </tr>
+                  ) : null}
+                </Fragment>
+              );
+            })}
+            {rows.length === 0 ? (
+              <tr>
+                <td colSpan={columns.length}>{emptyState}</td>
+              </tr>
+            ) : null}
+          </tbody>
+        </table>
+      </div>
+      <PaginationControls
+        count={rows.length}
+        currentPage={pagination.currentPage}
+        totalPages={pagination.totalPages}
+        startItem={pagination.startItem}
+        endItem={pagination.endItem}
+        pageSize={pageSize}
+        pageSizeOptions={pageSizeOptions}
+        onPageChange={onPageChange}
+        onPageSizeChange={onPageSizeChange}
+        t={t}
+      />
+    </MonitoringPanel>
+  );
+}
